rangelist: separate declaration and definition
